fixed the bug in converting HDF5 8-bit signed integer to DAP 16 bit integer. This bug is reported by a NASA user and you can test it with EffectiveTemperature on OMDOA3e OMI product.

+++ h5get.cc
@@ -194,7 +194,7 @@ string get_dap_type(hid_t type)
             if (sign == H5T_SGN_NONE)      
                 return BYTE;    
             else
-                return INT16;
+                return INT8;
         }
 
         if (size == 2) {
